Legend - refactor update, as differences were large
